feature: MPT-20733 add API route declarations
