+++ mpt_extension_sdk/__init__.py
@@ -1,10 +1,5 @@
 from mpt_extension_sdk.extension_app import ExtensionApp
-from mpt_extension_sdk.routing import APIRouter, EventRouter, PlugRouter, ScheduleRouter
 
 __all__ = [  # noqa: WPS410
-    "APIRouter",
-    "EventRouter",
     "ExtensionApp",
-    "PlugRouter",
-    "ScheduleRouter",
 ]
