+++ mpt_extension_sdk/api/builders/event.py
@@ -17,7 +17,7 @@ from mpt_extension_sdk.observability.tracing import (
     start_event_span,
 )
 from mpt_extension_sdk.pipeline import EventBaseContext, build_context
-from mpt_extension_sdk.routing.models import EventDeliveryMode, EventRouteDefinition
+from mpt_extension_sdk.routing import EventDeliveryMode, EventRouteCallback, EventRouteDefinition
 from mpt_extension_sdk.runtime.logging import set_event_context
 from mpt_extension_sdk.services.mpt_api_service.api_service import MPTAPIService
 from mpt_extension_sdk.services.mpt_api_service.task import TaskService
@@ -178,7 +178,7 @@ async def build_authenticated_context(
 
 
 async def run_handler(
-    event_handler: Callable[..., Awaitable[None] | None], event: Any, context: EventBaseContext
+    event_handler: EventRouteCallback, event: Any, context: EventBaseContext
 ) -> None:
     """Invoke a handler and await the result if it is a coroutine."""
     handler_result = event_handler(event, context)
