+++ mpt_extension_sdk/routing/__init__.py
@@ -1,32 +1,31 @@
+from mpt_extension_sdk.routing.enums import EventDeliveryMode, HTTPMethod, RouteType
 from mpt_extension_sdk.routing.models import (
     APIRouteDefinition,
     BaseRouteDefinition,
-    EventDeliveryMode,
     EventRouteDefinition,
     PlugRouteDefinition,
-    RouteCallback,
-    RouteType,
     ScheduleRouteDefinition,
 )
-from mpt_extension_sdk.routing.routers import (
-    APIRouter,
-    BaseExtensionRouter,
-    EventRouter,
-    PlugRouter,
-    ScheduleRouter,
-)
+from mpt_extension_sdk.routing.routers.api import APIRouter
+from mpt_extension_sdk.routing.routers.base import BaseExtensionRouter
+from mpt_extension_sdk.routing.routers.event import EventRouter
+from mpt_extension_sdk.routing.routers.plug import PlugRouter
+from mpt_extension_sdk.routing.routers.schedule import ScheduleRouter
+from mpt_extension_sdk.routing.types import APIRouteCallback, EventRouteCallback
 
 __all__ = [  # noqa: WPS410
+    "APIRouteCallback",
     "APIRouteDefinition",
     "APIRouter",
     "BaseExtensionRouter",
     "BaseRouteDefinition",
     "EventDeliveryMode",
+    "EventRouteCallback",
     "EventRouteDefinition",
     "EventRouter",
+    "HTTPMethod",
     "PlugRouteDefinition",
     "PlugRouter",
-    "RouteCallback",
     "RouteType",
     "ScheduleRouteDefinition",
     "ScheduleRouter",
