+++ mpt_extension_sdk/routing/types.py
@@ -1,0 +1,6 @@
+from collections.abc import Awaitable, Callable
+from typing import Any
+
+RouteCallback = Callable[..., Awaitable[Any] | Any]
+EventRouteCallback = Callable[..., Awaitable[None] | None]
+APIRouteCallback = Callable[..., Awaitable[object] | object]
