+++ mpt_extension_sdk/schemas.py
@@ -1,0 +1,5 @@
+from mpt_extension_sdk.models.base import BaseModel
+
+
+class BaseSchema(BaseModel):
+    """Base request-body schema for API routes."""
