+++ tests/routing/routers/conftest.py
@@ -1,0 +1,8 @@
+from collections.abc import Callable
+
+import pytest
+
+
+@pytest.fixture
+def route_handler(mocker):
+    return mocker.Mock(spec=Callable)
